Support user's password.
To support user passwords, we have to support users. Users have a
dedicated table in the database. We have to create, update or delete
users in this table depending of the request. A new plugin is then
introduced to deal with principals requests and to trigger specific
actions related to users. For instance, `MKCOL` of a new principal will
create a new user.

We use the `password_*` API, introduced in PHP5.5. Because we require at
least PHP5.4.7, we use a userland implementation. This API automatically
generate and deal with salt (aka realm). Thus, we can remove the realm
support in sabre/katana (i.e. removing the `authentication.realm` in the
configuration file, removing the realm in the authentication etc.).