Check if a new version exists.
From the server-side: We introduce a new node, called `/versions`, that
returns a JSON payload of kind

    {"current_version": …}

if no new version exists, or of kind

    {"current_version": …, "next_versions": […]}

if new versions exist.

From the client-side: We send a `GET` request on this new node and check
the `next_versions` property. If new versions exists, we prompt a
(dissmissable) message to the user saying that a new version is
available.