Add the `xml*ToPrimitive` methods.
The native XML objects are a little bit over-complicated for our usages.
So we introduce the `PrimiteXMLElement` object, that has a `name` (Clark
notation), `attributes`, `children` and `value` properties. Moreover, it
has a `toString` method (returning the `value` property).

Then, in the `multiStatus` parser, we are able to deeply parse a
response.

Moreover, we changed the `properties` property by `prop` in the result
of the `multiStatus` function, in order to mimic the real XML response.