Close opened streams after each mail sent.
Why? Because `Hoa\File` is based on `Hoa\Stream` and the latter is a
multiton. If we read from a stream like `http://….mapbox.com/…`, first
read will be ok but not the second one because the internal stream
pointer is at the end of the stream and this kind of stream is not
rewindable. So we close it manually instead of waiting it to be closed
at the end of the processus, in order to force to reopen for the next
mail.

Moreover, it saves memory.