Send the database errors on the client-side.
Instead of checking that a server-side check is strictly equal to
`false` to invalid a data, we check if it is different of `true`. If it
is not `true`, it can be different value that can be interpreted as an
error: Either `false` or an object containing an error object. In the
case of the database check, an error object is sent by the server to the
client containing an exception message. This is useful to provide a
feedback to the user.