Move the database name field after all the others.
Why? Because if the user chooses a database name before the username and
password, the “Cannot connect to the datase” error will be prompted to
her. This is not the expected workflow. Username and password are
required before the database name in order to ensure a proper connection
and avoid a useless error.